Fix compile errors on macOS.

+++ Dev/Cpp/src/SoundGodot/EffekseerGodot.SoundPlayer.cpp
@@ -20,7 +20,7 @@ Effekseer::SoundHandle SoundPlayer::Play(Effekseer::SoundTag tag, const Instance
 	auto data = (SoundData*)parameter.Data.Get();
 
 	godot::Dictionary args;
-	args["tag"] = reinterpret_cast<size_t>(tag);
+	args["tag"] = reinterpret_cast<int64_t>(tag);
 	args["emitter"] = reinterpret_cast<godot::Object*>(parameter.UserData);
 	args["stream"] = data->GetStream();
 	args["volume"] = parameter.Volume;
@@ -30,38 +30,38 @@ Effekseer::SoundHandle SoundPlayer::Play(Effekseer::SoundTag tag, const Instance
 	args["position"] = ToGdVector3(parameter.Position);
 	args["distance"] = parameter.Distance;
 
-	auto result = (size_t)soundContext_->call("play", args);
+	auto result = (size_t)(int64_t)soundContext_->call("play", args);
 	return reinterpret_cast<Effekseer::SoundHandle>((size_t)result);
 }
 
 void SoundPlayer::Stop(Effekseer::SoundHandle handle, Effekseer::SoundTag tag)
 {
-	soundContext_->call("stop", reinterpret_cast<size_t>(handle));
+	soundContext_->call("stop", reinterpret_cast<int64_t>(handle));
 }
 
 void SoundPlayer::Pause(Effekseer::SoundHandle handle, Effekseer::SoundTag tag, bool pause)
 {
-	soundContext_->call("pause", reinterpret_cast<size_t>(handle), pause);
+	soundContext_->call("pause", reinterpret_cast<int64_t>(handle), pause);
 }
 
 bool SoundPlayer::CheckPlaying(Effekseer::SoundHandle handle, Effekseer::SoundTag tag)
 {
-	return (bool)soundContext_->call("check_playing", reinterpret_cast<size_t>(handle));
+	return (bool)soundContext_->call("check_playing", reinterpret_cast<int64_t>(handle));
 }
 
 void SoundPlayer::StopTag(Effekseer::SoundTag tag)
 {
-	soundContext_->call("stop_tag", reinterpret_cast<size_t>(tag));
+	soundContext_->call("stop_tag", reinterpret_cast<int64_t>(tag));
 }
 
 void SoundPlayer::PauseTag(Effekseer::SoundTag tag, bool pause)
 {
-	soundContext_->call("pause_tag", reinterpret_cast<size_t>(tag), pause);
+	soundContext_->call("pause_tag", reinterpret_cast<int64_t>(tag), pause);
 }
 
 bool SoundPlayer::CheckPlayingTag(Effekseer::SoundTag tag)
 {
-	return (bool)soundContext_->call("check_playing_tag", reinterpret_cast<size_t>(tag));
+	return (bool)soundContext_->call("check_playing_tag", reinterpret_cast<int64_t>(tag));
 }
 
 void SoundPlayer::StopAll()
